[#noissue] fix test
fix integration test for grpc
